Create Customer_segment_transition_analysis.sql
# Customer Segment Transition Analysis

This query tracks the movement of customers across various segments over time. The focus is on understanding how customers upgrade, downgrade, remain in the same segment, churn out, or those recently gained.

The analysis is restricted to data from the years 2022 to 2024, 

The database contains the relationship of customer id and segment. A quarterly snapshot of this relationship is provided within the table. Therefore after pivoting the table it was possible to see how this segmentation has been changing over time.

- **Segment values** are defined based on the `dynamic_segmentation_engagement` attribute as follows:
  - `Digital` segments are marked as `1`.
  - `Mid` or `Medium` segments are marked as `2`.
  - `High` segments are marked as `3`.
  - `Scaled` segments are marked as `4`.

Finally the number of rows with a certain combination of transitions across time is counted. e.g. a transition could be those customers who have always been in the same category